oh-tab-czz: split useHalFlow into smaller hooks
