+++ src/webview-src/components/app/halFlow/blobToBase64.ts
@@ -1,0 +1,20 @@
+export function blobToBase64(blob: Blob): Promise<string> {
+  return new Promise((resolve, reject) => {
+    const reader = new FileReader();
+    reader.onload = () => {
+      const result = reader.result;
+      if (typeof result !== 'string') {
+        reject(new Error('Failed to read recorded audio.'));
+        return;
+      }
+      const comma = result.indexOf(',');
+      resolve(comma >= 0 ? result.slice(comma + 1) : result);
+    };
+    reader.onerror = () => {
+      const error = reader.error ?? new Error('Failed to read recorded audio.');
+      reject(error);
+    };
+    reader.readAsDataURL(blob);
+  });
+}
+
